+++ CHANGELOG.md
@@ -546,7 +546,7 @@ Bugfixes:
 Bugfixes:
 
 * Core: fixed LFI vulnerability.
-* Core: you can now override the template for sub navigation and the navigation: {$var|getnavigation:'page':{$page.id}:2:null:null:'/core/layout/templates/subnavigation.tpl'}
+* Core: you can now override the template for sub navigation and the navigation: {$var|getnavigation:'page':{$page.id}:2:null:null:'/Core/Layout/Templates/subnavigation.tpl'}
 * Extensions: installing a pre-uploaded theme from the themes overview now installs the selected theme instead of the last theme.
 * Mailmotor: fix CSV address imports.
 * Pages: include footer/meta subpages in the linkedlist.